Add authorization with Pundit

+++ app/controllers/answers_controller.rb
@@ -1,8 +1,7 @@
 class AnswersController < ApplicationController
 
   before_action :authenticate_user!
-  before_action :load_own_answer, only: [:update, :destroy]
-  before_action :load_answer, only: [:show, :make_best]
+  before_action :load_answer, only: [:show, :update, :destroy, :make_best]
   after_action :publish_answer, only: :create
   include Voted
 
@@ -14,6 +13,7 @@ class AnswersController < ApplicationController
   end
 
   def update
+    authorize @answer
     @answer.update(answer_params)
     respond_with @answer
   end
@@ -24,6 +24,7 @@ class AnswersController < ApplicationController
   end
 
   def destroy
+    authorize @answer
     respond_with @answer.destroy
   end
 
@@ -40,13 +41,6 @@ class AnswersController < ApplicationController
   def load_answer
     @answer = Answer.find params[:id]
     @question = @answer.question
-  end
-
-  def load_own_answer
-    load_answer
-    unless @answer.user == current_user
-      render text: 'You do not have permission to modify this answer.', status: 403
-    end
   end
 
   def answer_params
